# Conflicts:
#	src/app/Models/Classes/UserData.ts
#	src/app/Services/DataChangeService/data-change.service.ts
#	src/app/Services/DialogService/dialog.service.ts
#	src/app/SharedComponents/confirm-dialog/confirm-dialog.component.html
#	src/app/SharedComponents/list-element/list-element.component.ts
#	src/app/app-routing.module.ts
#	src/app/app.component.html
#	src/app/app.module.ts

+++ src/app/SharedComponents/edit-buchung/edit-buchung.component.ts
@@ -22,7 +22,7 @@ export class EditBuchungComponent implements OnInit {
   kategorien = computed<{ id: number, name: string }[]>(() => {
     this.dataService.updated();
     return this.dataProvider.getBuchungsKategorienMitEmpty();
-  }) ;
+  });
   isGeplanteBuchungChecked!: boolean;
 
   buchung = signal<IBuchung | undefined>(undefined);
@@ -89,7 +89,7 @@ export class EditBuchungComponent implements OnInit {
   }
 
   onBuchungsKategorieChanged(): void {
-    if(this.buchung()!.data.buchungsKategorie == -1) {
+    if (this.buchung()!.data.buchungsKategorie == -1) {
       this.dialogService.showBuchungsKategorienDialog();
       this.buchung()!.data.buchungsKategorie = 0;
     }
@@ -99,6 +99,7 @@ export class EditBuchungComponent implements OnInit {
   onGeplanteBuchungChange(newValue: boolean) {
     this.buchung()!.data.geplanteBuchung = newValue;
     this.isGeplanteBuchungChecked = newValue;
+    this.buchung()!.data.buchungsKategorie = 0;
     this.updateSaveButton();
   }
 
@@ -106,25 +107,25 @@ export class EditBuchungComponent implements OnInit {
     if (this.buchung()!.data.betrag !== 0 && this.buchung()!.data.betrag !== null) {
       if (this.isSaveButtonEnabled()) {
         let isBetragZuHoch = this.isBetragZuHoch();
-        if (!isBetragZuHoch || this.dataProvider.getMonthByDate(this.buchung()!.data.date).totalBudget! < 1) {
-          if(this.buchung()?.data.geplanteBuchung) {
-            if(this.oldBuchung?.data.geplanteBuchung) {
+        if (!(isBetragZuHoch && this.hasBetragChanged()) || this.dataProvider.getMonthByDate(this.buchung()!.data.date).totalBudget! < 1) {
+          if (this.buchung()?.data.geplanteBuchung) {
+            if (this.oldBuchung?.data.geplanteBuchung) {
               this.dataChangeService.editGeplanteAusgabeBuchung(this.buchung()!)
             } else {
               this.dataChangeService.addGeplanteAusgabeBuchung(this.buchung()!.data!)
               this.dataChangeService.deleteBuchung(this.buchung()!.id)
             }
           } else {
-            if(!this.oldBuchung?.data.geplanteBuchung) {
+            if (!this.oldBuchung?.data.geplanteBuchung) {
               this.dataChangeService.editBuchung(this.buchung()!)
             } else {
               this.dataChangeService.addBuchung(this.buchung()!.data!)
               this.dataChangeService.deleteGeplanteAusgabeBuchung(this.buchung()!.id)
             }
           }
-          this.router.navigate(['/']);
+          this.router.navigate(['home']);
         } else {
-          if(this.dataProvider.getSettings().toHighBuchungenEnabled) {
+          if (this.dataProvider.getSettings().toHighBuchungenEnabled) {
             const confirmDialogViewModel: ConfirmDialogViewModel = {
               title: 'Betrag ist zu hoch',
               message: `Der Betrag überschreitet dein Budget für ${this.buchung()!.data.date.toLocaleDateString() === new Date().toLocaleDateString() ? 'heute' : 'den ' + this.buchung()!.data.date.toLocaleDateString()}. Trotzdem fortfahren?`,
@@ -132,15 +133,15 @@ export class EditBuchungComponent implements OnInit {
                 this.dialogService.isConfirmDialogVisible = false;
               },
               onConfirmClicked: () => {
-                if(this.buchung()?.data.geplanteBuchung) {
-                  if(this.oldBuchung?.data.geplanteBuchung) {
+                if (this.buchung()?.data.geplanteBuchung) {
+                  if (this.oldBuchung?.data.geplanteBuchung) {
                     this.dataChangeService.editGeplanteAusgabeBuchung(this.buchung()!)
                   } else {
                     this.dataChangeService.addGeplanteAusgabeBuchung(this.buchung()!.data!)
                     this.dataChangeService.deleteBuchung(this.buchung()!.id)
                   }
                 } else {
-                  if(!this.oldBuchung?.data.geplanteBuchung) {
+                  if (!this.oldBuchung?.data.geplanteBuchung) {
                     this.dataChangeService.editBuchung(this.buchung()!)
                   } else {
                     this.dataChangeService.addBuchung(this.buchung()!.data!)
@@ -148,7 +149,7 @@ export class EditBuchungComponent implements OnInit {
                   }
                 }
                 this.dialogService.isConfirmDialogVisible = false;
-                this.router.navigate(['/']);
+                this.router.navigate(['home']);
               }
             }
             this.dialogService.showConfirmDialog(confirmDialogViewModel);
@@ -165,13 +166,18 @@ export class EditBuchungComponent implements OnInit {
     }
   }
 
-  isBetragZuHoch() {
-    if(this.buchung()?.data.buchungsKategorie == -1) {
-      return !this.availableMoneyCapped().noData && this.getRestGeplantAusgabenSumme();
+  isBetragZuHochUngeplant() {
+    if (this.buchung()?.data.buchungsKategorie == -1) {
+      if (this.availableMoneyCapped().noData) {
+        return false;
+      }
+      return this.getRestGeplantAusgabenSumme();
     } else {
-      return !this.availableMoneyCapped().noData && this.getAvailableMoneyDay()! < 0
+      if (this.availableMoneyCapped().noData) {
+        return false;
+      }
+      return this.getAvailableMoneyDay()! === -1
     }
-
   }
 
   getRestGeplantAusgabenSumme() {
@@ -189,6 +195,13 @@ export class EditBuchungComponent implements OnInit {
     this.executeExitAction();
   }
 
+  private isBetragZuHoch() {
+    if (this.buchung() && this.oldBuchung && this.buchung()!.data.geplanteBuchung) {
+      return this.buchung!()!.data.betrag! > (this.availableMonayForGeplanteAusgabenKategorien().restgeldBetrag ?? 0) + (this.oldBuchung.data!.betrag ?? 0);
+    }
+    return this.isBetragZuHochUngeplant();
+  }
+
   onDateChange() {
     if (this.date)
       this.buchung()!.data.date = new Date(this.date);
@@ -198,10 +211,7 @@ export class EditBuchungComponent implements OnInit {
   }
 
   onTimeChange(event: any) {
-    const [hours, minutes] = event.target.value.split(':');
-    const date = new Date();
-    date.setHours(+hours, +minutes);
-    this.buchung()!.data.time = date.toLocaleTimeString('de-DE', {hour: '2-digit', minute: '2-digit'});
+    this.buchung()!.data.time = event;
     this.updateSaveButton();
   }
 
@@ -234,21 +244,12 @@ export class EditBuchungComponent implements OnInit {
   }
 
   private hasBuchungChanged() {
-    console.log(2)
-    const x = !(this.buchung()!.data.geplanteBuchung === this.oldBuchung?.data.geplanteBuchung && this.buchung()!.data.buchungsKategorie === this.oldBuchung?.data.buchungsKategorie && this.buchung()!.data.betrag === this.oldBuchung?.data.betrag && this.buchung()!.data.title === this.oldBuchung?.data.title && this.buchung()!.data.beschreibung === this.oldBuchung?.data.beschreibung && this.buchung()!.data.date.getDate() === this.oldBuchung.data.date.getDate() && this.buchung()!.data.time === this.oldBuchung.data.time)
-    console.log(x)
+    const x = !(this.buchung()!.data.geplanteBuchung === this.oldBuchung?.data.geplanteBuchung && this.buchung()!.data.buchungsKategorie === this.oldBuchung?.data.buchungsKategorie && this.buchung()!.data.betrag === this.oldBuchung?.data.betrag && this.buchung()!.data.title === this.oldBuchung?.data.title && this.buchung()!.data.beschreibung === this.oldBuchung?.data.beschreibung && new Date(this.buchung()!.data.date).getDate() === new Date(this.oldBuchung.data.date).getDate() && this.buchung()!.data.time === this.oldBuchung.data.time)
     return x;
   }
 
   private isSaveAble() {
-    console.log(1)
-    console.log(this.oldBuchung)
-    console.log(this.buchung())
-    console.log(this.buchung()?.data.betrag)
-
-    const x = (this.buchung()?.data.betrag !== null && this.buchung()?.data.betrag !== 0) && this.hasBuchungChanged();
-    console.log(x)
-    return x;
+    return (this.buchung()?.data.betrag !== null && this.buchung()?.data.betrag !== 0) && this.hasBuchungChanged() && !(this.buchung()?.data.geplanteBuchung && this.buchung()?.data.buchungsKategorie === 0);
   }
 
   private executeExitAction() {
@@ -290,6 +291,10 @@ export class EditBuchungComponent implements OnInit {
 
   protected getAvailableMoneyDay() {
     const x = this.ut.toFixedDown(this.availableMoney().availableForDayIst! + this.oldBuchung?.data.betrag! - this.buchung()?.data.betrag!, 2);
-    return x && x > 0 ? x : 0;
+    return x && x < 0 ? -1 : x;
+  }
+
+  protected hasBetragChanged() {
+    return this.buchung()?.data.betrag !== this.oldBuchung?.data.betrag;
   }
 }
